@@ -20,88 +20,99 @@
   "license": "MIT",
   "dependencies": {
     "bcrypt-nodejs": "0.0.3",
-    "body-parser": "1.15.0",
-    "bootstrap": "3.3.6",
-    "classnames": "2.2.3",
-    "cookie-parser": "1.4.1",
-    "connect-mongo": "1.1.0",
+    "body-parser": "1.15.2",
+    "bootstrap": "3.3.7",
+    "classnames": "2.2.5",
+    "connect-ensure-login": "0.1.1",
+    "connect-mongo": "1.3.2",
+    "connect-redis": "3.1.0",
+    "cookie-parser": "1.4.3",
+    "csurf": "1.9.0",
     "cuid": "1.3.8",
-    "debug": "2.2.0",
     "escape-string-regexp": "1.0.5",
     "express": "4.13.4",
-    "express-session": "1.13.0",
-    "font-awesome": "4.5.0",
-    "history": "2.0.0",
-    "jquery": "2.2.1",
+    "express-hbs": "1.0.3",
+    "express-session": "1.14.0",
+    "font-awesome": "4.6.3",
+    "history": "3.0.0",
+    "ip-address": "5.8.2",
+    "jquery": "3.1.0",
     "lackey-mongoose-version": "1.0.2",
-    "lodash": "4.5.1",
-    "moment": "2.11.2",
-    "moment-range": "2.1.0",
-    "moment-timezone": "0.5.0",
-    "mongoose": "4.4.5",
+    "lodash": "4.14.1",
+    "moment-config-trejgun": "git://github.com/TrejGun/moment-config-trejgun.git#v0.0.6",
+    "mongoose": "4.6.0",
     "morgan": "1.7.0",
-    "nodemailer": "2.2.1",
-    "nodemailer-mailgun-transport": "1.0.1",
+    "nodemailer": "2.3.2",
+    "nodemailer-mailgun-transport": "1.2.1",
+    "oauth2orize": "1.5.0",
     "passport": "0.3.2",
     "passport-local": "1.0.0",
     "passport-facebook": "1.0.3",
     "passport-google-oauth": "0.1.5",
+    "passport-oauth2": "1.3.0",
+    "passport-http": "0.3.0",
+    "passport-http-bearer": "1.0.1",
+    "passport-oauth2-client-password": "0.1.2",
     "q": "1.4.1",
-    "react": "0.14.7",
-    "react-bootstrap": "0.28.3",
-    "react-dom": "0.14.7",
-    "react-google-maps": "4.9.1",
-    "react-redux": "4.4.0",
-    "react-router": "2.0.0",
-    "react-router-bootstrap": "0.20.1",
-    "redux": "3.3.1",
-    "react-router-redux": "4.0.0",
-    "redux-logger": "2.6.0",
-    "redux-thunk": "1.0.3",
+    "query-types": "0.1.4",
+    "react": "15.3.0",
+    "react-bootstrap": "0.30.1",
+    "react-dom": "15.3.0",
+    "react-google-maps": "4.11.0",
+    "react-redux": "4.4.5",
+    "react-router": "2.6.1",
+    "react-router-bootstrap": "0.23.1",
+    "react-router-redux": "4.0.5",
+    "redux": "3.5.2",
+    "redis": "2.4.2",
+    "redux-logger": "2.6.1",
+    "redux-thunk": "2.1.0",
     "twilio": "3.3.0-edge",
     "twit": "2.1.3",
-    "zxcvbn": "4.2.0"
+    "winston": "2.2.0",
+    "zxcvbn": "4.3.0"
   },
   "devDependencies": {
-    "babel-cli": "6.5.1",
-    "babel-core": "6.5.2",
-    "babel-loader": "6.2.4",
-    "babel-plugin-espower": "2.1.2",
-    "babel-plugin-react-transform": "2.0.0",
-    "babel-plugin-transform-decorators-legacy": "1.3.4",
-    "babel-plugin-transform-runtime": "6.5.2",
-    "babel-preset-es2015": "6.5.0",
-    "babel-preset-react": "6.5.0",
-    "babel-preset-stage-0": "6.5.0",
-    "babel-runtime": "6.5.0",
-    "core-decorators": "0.11.0",
+    "async": "2.0.1",
+    "babel-config-trejgun": "git://github.com/TrejGun/babel-config-trejgun.git#v0.0.12",
+    "core-decorators": "0.12.0",
     "css-loader": "0.23.1",
-    "eslint-config-trejgun": "git://github.com/TrejGun/eslint-config-trejgun.git#0.0.2",
-    "extract-text-webpack-plugin": "git://github.com/TrejGun/extract-text-webpack-plugin.git#4c2af8a819b30e1cd6d23e329b7492f808ee6f99",
-    "file-loader": "0.8.5",
-    "istanbul": "^1.0.0-alpha",
-    "less": "2.6.0",
-    "less-loader": "2.2.2",
-    "mocha": "2.4.5",
-    "nodemon": "1.9.1",
-    "power-assert": "1.2.0",
-    "react-transform-hmr": "1.0.2",
-    "style-loader": "0.13.0",
-    "webpack": "1.12.14",
-    "webpack-dev-middleware": "1.5.1",
-    "webpack-hot-middleware": "2.7.1"
+    "eslint-config-trejgun": "git://github.com/TrejGun/eslint-config-trejgun.git#v0.0.4",
+    "extract-text-webpack-plugin": "2.0.0-beta.3",
+    "file-loader": "0.9.0",
+    "istanbul": "1.1.0-alpha.1",
+    "less": "2.7.1",
+    "less-loader": "2.2.3",
+    "json-loader": "0.5.4",
+    "mocha": "3.0.0",
+    "nodemon": "1.10.0",
+    "passport-stub": "1.1.1",
+    "power-assert": "1.4.1",
+    "react-transform-hmr": "1.0.4",
+    "rimraf": "2.5.4",
+    "style-loader": "0.13.1",
+    "supertest": "2.0.0",
+    "supertest-as-promised": "4.0.0",
+    "webpack": "2.1.0-beta.20",
+    "webpack-dev-middleware": "1.6.1",
+    "webpack-hot-middleware": "2.12.2"
   },
   "engines": {
-    "node": "v5.3.0"
+    "node": ">=6.0.0"
   },
   "scripts": {
+    "postinstall": "rimraf ./build && babel -d ./build/server ./server --copy-files && babel -d ./build/client ./client --ignore build --copy-files",
     "build": "webpack -c --config server/configs/webpack.js",
     "db": "NODE_ENV=${NODE_ENV:=test} babel-node tools/db-category.js",
-    "coverage": "NODE_ENV=${NODE_ENV:=test} babel-node node_modules/istanbul/lib/cli cover --report text --report html node_modules/.bin/_mocha test/controllers/abstract",
-    "lint": "eslint server/* client/assets/js/* test/*",
-    "prod": "NODE_ENV=${NODE_ENV:=production} babel-node server/server.js",
-    "start": "NODE_ENV=${NODE_ENV:=development} TWITTER_API=${TWITTER_API:=true} DEBUG=web:* nodemon server/server.js",
-    "test": "NODE_ENV=${NODE_ENV:=test} babel-node node_modules/.bin/_mocha",
-    "travis": "NODE_ENV=${NODE_ENV:=test} babel-node node_modules/istanbul/lib/cli cover --report text node_modules/.bin/_mocha test/controllers/abstract"
+    "lint": "eslint source/* test/* --quiet --fix",
+    "port": "lsof -t -i tcp:8080 | xargs kill",
+    "prod": "NODE_ENV=${NODE_ENV:=production} babel-node source/server.js",
+    "start:pm2": "pm2 start ecosystem.json --env dev",
+    "start:nodemon": "NODE_ENV=${NODE_ENV:=development} TWITTER_API=${TWITTER_API:=true} DEBUG=web:* nodemon server/blog.js",
+    "stop": "pm2 stop --watch all; pm2 delete all;",
+    "test": "NODE_ENV=${NODE_ENV:=test} ABL_DEBUG=${ABL_DEBUG:=info} babel-node node_modules/.bin/_mocha",
+    "manual": "NODE_ENV=${NODE_ENV:=test} ABL_DEBUG=${ABL_DEBUG:=info} node_modules/.bin/babel-node",
+    "coverage": "NODE_ENV=${NODE_ENV:=test} node_modules/.bin/babel-node node_modules/istanbul/lib/cli cover node_modules/.bin/_mocha test/routes/* test/connect test/utils test/controllers/*",
+    "travis": "babel-node node_modules/istanbul/lib/cli cover node_modules/.bin/_mocha test/routes/* test/connect test/utils test/controllers/*"
   }
 }
