@@ -1,6 +1,5 @@
-"use strict";
-
 import React, {PropTypes, Component} from "react";
+
 
 export default class EML extends Component {
 
