@@ -1,12 +1,14 @@
-"use strict";
-
 import React, {PropTypes, Component} from "react";
 import {connect} from "react-redux";
 import API from "../../../utils/API";
 import {bindActionCreators} from "redux";
-import {Input, ButtonInput, Pagination, ListGroup, ListGroupItem, Label} from "react-bootstrap";
+import {Col, FormGroup, FormControl, Button, ListGroup, ListGroupItem, Pagination} from "react-bootstrap";
 import {Link} from "react-router";
+import {LinkContainer} from "react-router-bootstrap";
+
+// import {Link} from "react-router";
 import {UPDATE_USER, UPDATE_USER_LIST} from "../../../constants/constants";
+
 
 const userUpdate = data =>
 	dispatch =>
@@ -123,34 +125,22 @@ export default class UserList extends Component {
 
 	showForm() {
 		return (
-			<form onSubmit={::this.onSubmit}>
-				<div className="row">
-					<div className="col-sm-4">
-						<Input
-							id="name-search"
-							autoFocus="true"
-							type="text"
+			<form onSubmit={::this.onSubmit} className="form-horizontal">
+				<FormGroup controlId="formHorizontalEmail">
+					<Col sm={4}>
+						<FormControl
+							type="email"
+							name="email"
 							value={this.state.email}
-							placeholder="User email"
-							onChange={(e) => this.setState({email: e.target.value})}
+							placeholder="me@example.com"
+							onChange={e => this.setState({email: e.target.value})}
 						/>
-					</div>
-					<div className="col-sm-1">
-						<ButtonInput
-							type="submit"
-							value="Search"
-							bsStyle="info"
-							disabled={this.validateEmailLength()}
-						/>
-					</div>
-					<div className="col-sm-2">
-						<h5 className="text-right">Per page:</h5>
-					</div>
-					<div className="col-sm-2">
-						<Input
-							id="limit"
-							type="select"
-							defaultValue={this.state.limit}
+					</Col>
+
+					<Col sm={2}>
+						<FormControl
+							componentClass="select"
+							placeholder="select"
 							onChange={(e) => this.onChangeLimit(~~e.target.value)}
 						>
 							<option value="5">5</option>
@@ -158,10 +148,18 @@ export default class UserList extends Component {
 							<option value="20">20</option>
 							<option value="50">50</option>
 							<option value="100">100</option>
-						</Input>
-					</div>
-				</div>
-				{::this.showPagination()}
+						</FormControl>
+					</Col>
+
+					<Col sm={1}>
+						<Button
+							type="submit"
+							disabled={this.state.disabled}
+						>
+							Search
+						</Button>
+					</Col>
+				</FormGroup>
 			</form>
 		);
 	}
@@ -172,21 +170,10 @@ export default class UserList extends Component {
 				<div className="row">
 					<div className="col-sm-8">
 						<h4 className="list-group-item-heading">
-							<Link to={"/admin/user/edit/" + user._id}>
-								{`${user.firstName} ${user.lastName}`}
+							<Link to={`/admin/user/edit/${user._id}`}>
+								{`${user.fullName}`}
 							</Link>
 						</h4>
-						Email: {user.email}<br/>
-
-						{user.role === "admin" ?
-							<Label bsStyle="success">Admin</Label>
-							: <Label bsStyle="primary">User</Label>}{" "}
-						{!user.isActive ?
-							<Label bsStyle="default">Not active</Label>
-							: null}{" "}
-						{!user.isEmailVerified ?
-							<Label bsStyle="warning">Not verified</Label>
-							: null}{" "}
 					</div>
 				</div>
 			</ListGroupItem>
@@ -195,11 +182,9 @@ export default class UserList extends Component {
 
 	render() {
 		return (
-			<div className="row">
-				<div className="col-sm-8">
-					{this.showForm()}
-					{this.showList()}
-				</div>
+			<div className="container">
+				{this.showForm()}
+				{this.showList()}
 			</div>
 		);
 	}
