@@ -1,12 +1,10 @@
-"use strict";
-
 import url from "url";
 import {connect} from "react-redux";
 import React, {PropTypes, Component} from "react";
 import configs from "../../../../../server/configs/config";
 
-const config = configs[process.env.NODE_ENV];
 
+const config = configs[process.env.NODE_ENV];
 
 @connect(
 	state => ({
