@@ -1,6 +1,5 @@
-"use strict";
-
 import React, {Component} from "react";
+
 
 export default class Pagination extends Component {
 
