@@ -1,7 +1,6 @@
-"use strict";
-
 import React, {Component} from "react";
 import Title from "../partials/title";
+
 
 export default class Welcome extends Component {
 
