@@ -1,8 +1,7 @@
-"use strict";
-
 import React from "react";
 import TwitterList from "./twitterlist";
 import TwitterForm from "./twitterform";
+
 
 export default class TwitSearch extends React.Component {
 
