@@ -1,9 +1,7 @@
-"use strict";
-
 import React, {PropTypes, Component} from "react";
 import {bindActionCreators} from "redux";
 import {connect} from "react-redux";
-import {Input, ButtonInput} from "react-bootstrap";
+import {Col, FormGroup, ControlLabel, FormControl, Button} from "react-bootstrap";
 import API from "../../utils/API";
 import {email, password} from "../../../../../server/utils/constants/misc";
 import {Link} from "react-router";
@@ -90,7 +88,7 @@ export default class Login extends Component {
 			const r = 400;
 			const i = (window.innerHeight - r) / 2;
 			const s = (window.innerWidth - n) / 2;
-			const popup = window.open(link, "authorization", "height=" + r + ",width=" + n + ",top=" + i + ",left=" + s);
+			const popup = window.open(link, "authorization", `height=${r},width=${n},top=${i},left=${s}`);
 			if (window.focus) {
 				popup.focus();
 			}
@@ -109,36 +107,57 @@ export default class Login extends Component {
 
 					[<a href="#" onClick={this.open("/auth/google")}>google</a>]
 					[<a href="#" onClick={this.open("/auth/facebook")}>facebook</a>]
+					[<a href="#" onClick={this.open("/auth/goat")}>goat</a>]
 
 					[<Link to="/user/forgot">Forgot password?</Link>]
 
+					<br/>
+
 					<form className="form-horizontal" onSubmit={::this.onSubmit} autoComplete="off">
-						<Input
-							type="email"
-							name="email"
-							value={this.state.email}
-							placeholder="me@example.com"
-							label="Email"
-							wrapperClassName="col-xs-10"
-							labelClassName="col-sm-2"
-							onChange={e => this.setState({email: e.target.value})}
-						/>
-						<Input
-							type="password"
-							name="password"
-							value={this.state.password}
-							placeholder="******"
-							label="Password"
-							wrapperClassName="col-xs-10"
-							labelClassName="col-sm-2"
-							onChange={e => this.setState({password: e.target.value})}
-						/>
-						<ButtonInput
-							type="submit"
-							value="Login"
-							wrapperClassName="col-sm-offset-2 col-sm-10"
-							disabled={this.state.disabled}
-						/>
+						<FormGroup
+							controlId="formHorizontalEmail"
+							// validationState={this.getValidationState()}
+						>
+							<Col componentClass={ControlLabel} sm={2}>
+								Email
+							</Col>
+							<Col sm={10}>
+								<FormControl
+									type="email"
+									name="email"
+									value={this.state.email}
+									placeholder="me@example.com"
+									onChange={e => this.setState({email: e.target.value})}
+								/>
+							</Col>
+						</FormGroup>
+						<FormGroup
+							controlId="formHorizontalPassword"
+							// validationState={this.getValidationState()}
+						>
+							<Col componentClass={ControlLabel} sm={2}>
+								Password
+							</Col>
+							<Col sm={10}>
+								<FormControl
+									type="password"
+									name="password"
+									value={this.state.password}
+									placeholder="******"
+									onChange={e => this.setState({password: e.target.value})}
+								/>
+							</Col>
+						</FormGroup>
+						<FormGroup>
+							<Col smOffset={2} sm={10}>
+								<Button
+									type="submit"
+									disabled={this.state.disabled}
+								>
+									Login
+								</Button>
+							</Col>
+						</FormGroup>
 					</form>
 				</div>
 			</div>
