@@ -1,7 +1,6 @@
-"use strict";
-
 import React, {Component} from "react";
 import {Link} from "react-router";
+
 
 export default class Profile extends Component {
 
