@@ -1,14 +1,18 @@
-"use strict";
-
 import React, {PropTypes, Component} from "react";
 import {bindActionCreators} from "redux";
 import {connect} from "react-redux";
-import {Input, ButtonInput} from "react-bootstrap";
+import {Col, FormGroup, ControlLabel, FormControl, Button} from "react-bootstrap";
 import zxcvbn from "zxcvbn";
 import API from "../../utils/API";
 import {reEmail} from "../../../../../server/utils/constants/regexp";
-import {email, password, confirm, firstName, lastName/* , phoneNumber*/} from "../../../../../server/utils/constants/misc";
+import {
+	email,
+	password,
+	confirm,
+	fullName/* , phoneNumber*/
+} from "../../../../../server/utils/constants/misc";
 import {USER_LOGIN} from "../../constants/constants";
+
 
 const register = (data) =>
 	dispatch =>
@@ -33,8 +37,7 @@ export default class Register extends Component {
 		// phoneNumber: PropTypes.string,
 		password: PropTypes.string,
 		confirm: PropTypes.string,
-		firstName: PropTypes.string,
-		lastName: PropTypes.string,
+		fullName: PropTypes.string,
 		history: React.PropTypes.object,
 		register: PropTypes.func
 	};
@@ -48,8 +51,7 @@ export default class Register extends Component {
 		// phoneNumber
 		password,
 		confirm,
-		firstName,
-		lastName
+		fullName
 	};
 
 	state = {
@@ -57,8 +59,7 @@ export default class Register extends Component {
 		// phoneNumber: this.props.phoneNumber,
 		password: this.props.password,
 		confirm: this.props.confirm,
-		firstName: this.props.firstName,
-		lastName: this.props.lastName
+		fullName: this.props.lastName
 	};
 
 	onSubmit(e) {
@@ -69,85 +70,99 @@ export default class Register extends Component {
 			});
 	}
 
+	validateEmail() {
+		return reEmail.test(this.state.email) ? "success" : "error";
+	}
+
+	validatePassword() {
+		return zxcvbn(this.state.password).score >= 1 ? "success" : "error";
+	}
+
+	validateConfirm() {
+		return this.state.password === this.state.confirm ? "success" : "error";
+	}
+
 	render() {
 		return (
 			<div className="panel panel-default">
 				<div className="panel-body">
 					<form className="form-horizontal" onSubmit={::this.onSubmit} autoComplete="off">
-						<Input
-							type="email"
-							name="email"
-							value={this.state.email}
-							placeholder="me@example.com"
-							label="Email"
-							bsStyle={reEmail.test(this.state.email) ? null : "error"}
-							hasFeedback
-							wrapperClassName="col-xs-10"
-							labelClassName="col-sm-2"
-							onChange={e => this.setState({email: e.target.value})}
-						/>
-						{/*
-						<Input
-							type="text"
-							name="phoneNumber"
-							value={this.state.phoneNumber}
-							placeholder="+1234567890"
-							label="Phone number"
-							wrapperClassName="col-xs-10"
-							labelClassName="col-sm-2"
-							onChange={e => this.setState({phoneNumber: e.target.value})}
-						/>
-						*/}
-						<Input
-							type="password"
-							name="password"
-							value={this.state.password}
-							placeholder="******"
-							label="Password"
-							bsStyle={zxcvbn(this.state.password).score >= 1 ? null : "error"}
-							hasFeedback
-							wrapperClassName="col-xs-10"
-							labelClassName="col-sm-2"
-							onChange={e => this.setState({password: e.target.value})}
-						/>
-						<Input
-							type="password"
-							name="confirm"
-							value={this.state.confirm}
-							placeholder="******"
-							label="Confirm password"
-							bsStyle={this.state.password === this.state.confirm ? null : "error"}
-							hasFeedback
-							wrapperClassName="col-xs-10"
-							labelClassName="col-sm-2"
-							onChange={e => this.setState({confirm: e.target.value})}
-						/>
-						<Input
-							type="text"
-							name="firstName"
-							value={this.state.firstName}
-							placeholder="Fred"
-							label="First name"
-							wrapperClassName="col-xs-10"
-							labelClassName="col-sm-2"
-							onChange={e => this.setState({firstName: e.target.value})}
-						/>
-						<Input
-							type="text"
-							name="lastName"
-							value={this.state.lastName}
-							placeholder="Flintstone"
-							label="Last name"
-							wrapperClassName="col-xs-10"
-							labelClassName="col-sm-2"
-							onChange={e => this.setState({lastName: e.target.value})}
-						/>
-						<ButtonInput
-							type="submit"
-							value="Register"
-							wrapperClassName="col-sm-offset-2 col-sm-10"
-							disabled={this.state.disabled}
-						/>
+						<FormGroup
+							controlId="formHorizontalFirstName"
+						>
+							<Col componentClass={ControlLabel} sm={2}>
+								First name
+							</Col>
+							<Col sm={10}>
+								<FormControl
+									type="text"
+									name="fullName"
+									value={this.state.fullName}
+									placeholder="Gordon Freeman"
+									onChange={e => this.setState({fullName: e.target.value})}
+								/>
+							</Col>
+						</FormGroup>
+						<FormGroup
+							controlId="formHorizontalEmail"
+							validationState={this.validateEmail()}
+						>
+							<Col componentClass={ControlLabel} sm={2}>
+								Email
+							</Col>
+							<Col sm={10}>
+								<FormControl
+									type="email"
+									name="email"
+									value={this.state.email}
+									placeholder="me@example.com"
+									onChange={e => this.setState({email: e.target.value})}
+								/>
+							</Col>
+						</FormGroup>
+						<FormGroup
+							controlId="formHorizontalPassword"
+							validationState={this.validatePassword()}
+						>
+							<Col componentClass={ControlLabel} sm={2}>
+								Password
+							</Col>
+							<Col sm={10}>
+								<FormControl
+									type="password"
+									name="password"
+									value={this.state.password}
+									placeholder="******"
+									onChange={e => this.setState({password: e.target.value})}
+								/>
+							</Col>
+						</FormGroup>
+						<FormGroup
+							controlId="formHorizontalConfirm"
+							validationState={this.validateConfirm()}
+						>
+							<Col componentClass={ControlLabel} sm={2}>
+								Confirm password
+							</Col>
+							<Col sm={10}>
+								<FormControl
+									type="password"
+									name="confirm"
+									value={this.state.confirm}
+									placeholder="******"
+									onChange={e => this.setState({confirm: e.target.value})}
+								/>
+							</Col>
+						</FormGroup>
+						<FormGroup>
+							<Col smOffset={2} sm={10}>
+								<Button
+									type="submit"
+								>
+									Register
+								</Button>
+							</Col>
+						</FormGroup>
 					</form>
 				</div>
 			</div>
