@@ -1,12 +1,30 @@
-"use strict";
-
 export default {
-	test: {
+	[process.env.NODE_ENV]: {
 		realm: "user",
 		rendering: "server",
 		mongo: {
+			blog: {
+				url: "mongodb://localhost/goat-blog",
+				options: {
+					server: {
+						socketOptions: {
+							keepalive: 1
+						}
+					}
+				}
+			},
 			user: {
-				url: "mongodb://localhost/goat_test",
+				url: "mongodb://localhost/goat-user",
+				options: {
+					server: {
+						socketOptions: {
+							keepalive: 1
+						}
+					}
+				}
+			},
+			mail: {
+				url: "mongodb://localhost/goat-mail",
 				options: {
 					server: {
 						socketOptions: {
@@ -15,6 +33,10 @@ export default {
 					}
 				}
 			}
+		},
+		redis: {
+			port: 6379,
+			host: "localhost"
 		},
 		strategies: {
 			local: {
@@ -30,6 +52,14 @@ export default {
 				clientID: "",
 				clientSecret: "",
 				callbackURL: "http://localhost:3000/auth/google/callback"
+			},
+			abl: {
+				clientID: "goat-test-server",
+				clientSecret: "01123581321345589144233377610",
+				callbackURL: "http://localhost:9000/auth/goat/callback",
+				authorizationURL: "http://localhost:3000/dialog/authorize",
+				tokenURL: "http://localhost:3000/oauth/token",
+				profileURL: "http://localhost:3000/api/userinfo"
 			}
 		},
 		session: {
@@ -48,97 +78,25 @@ export default {
 			}
 		},
 		server: {
-			self: {
-				protocol: "http",
-				hostname: "localhost",
-				port: 3001
-			},
-			mailgun: {
-				api_key: "", // eslint-disable-line camelcase
-				domain: "",
-				from: "G.O.A.T. <no-reply@goat.com>"
-			},
-			twilio: {
-				AccountSID: "",
-				AuthToken: "",
-				from: "+15005550006" // https://www.twilio.com/docs/api/rest/test-credentials
-			},
-			twitter: {
-				consumer_key: "", // eslint-disable-line camelcase
-				consumer_secret: "", // eslint-disable-line camelcase
-				access_token: "", // eslint-disable-line camelcase
-				access_token_secret: "" // eslint-disable-line camelcase
-			}
-		}
-	},
-	development: {
-		realm: "user",
-		rendering: "server",
-		mongo: {
-			user: {
-				url: "mongodb://localhost/goat_dev",
-				options: {
-					server: {
-						socketOptions: {
-							keepalive: 1
-						}
-					}
-				}
-			}
+			protocol: "http",
+			hostname: "localhost",
+			port: 8080
 		},
-		strategies: {
-			local: {
-				usernameField: "email",
-				passwordField: "password"
-			},
-			facebook: {
-				clientID: "",
-				clientSecret: "",
-				callbackURL: "http://localhost:3000/auth/facebook/callback"
-			},
-			google: {
-				clientID: "",
-				clientSecret: "",
-				callbackURL: "http://localhost:3000/auth/google/callback"
-			}
+		mailgun: {
+			api_key: "", // eslint-disable-line camelcase
+			domain: "",
+			from: "G.O.A.T. <no-reply@goat.com>"
 		},
-		session: {
-			proxy: false,
-			secret: "keyboard_cat",
-			saveUninitialized: true,
-			resave: false,
-			rolling: true,
-			name: "goat.id",
-			cookie: {
-				path: "/",
-				httpOnly: true,
-				secure: false,
-				maxAge: 24 * 60 * 60 * 1000,
-				signed: false
-			}
+		twilio: {
+			AccountSID: "",
+			AuthToken: "",
+			from: "+15005550006" // https://www.twilio.com/docs/api/rest/test-credentials
 		},
-		server: {
-			self: {
-				protocol: "http",
-				hostname: "localhost",
-				port: 3001
-			},
-			mailgun: {
-				api_key: "", // eslint-disable-line camelcase
-				domain: "",
-				from: "G.O.A.T. <no-reply@goat.com>"
-			},
-			twilio: {
-				AccountSID: "",
-				AuthToken: "",
-				from: ""
-			},
-			twitter: {
-				consumer_key: "", // eslint-disable-line camelcase
-				consumer_secret: "", // eslint-disable-line camelcase
-				access_token: "", // eslint-disable-line camelcase
-				access_token_secret: "" // eslint-disable-line camelcase
-			}
+		twitter: {
+			consumer_key: "", // eslint-disable-line camelcase
+			consumer_secret: "", // eslint-disable-line camelcase
+			access_token: "", // eslint-disable-line camelcase
+			access_token_secret: "" // eslint-disable-line camelcase
 		}
 	}
 };
