@@ -1,26 +1,23 @@
-"use strict";
-
-import debug from "debug";
 import passport from "passport";
 import {Strategy as LocalStrategy} from "passport-local";
-import {Strategy as FacebookStrategy} from "passport-facebook";
 import {OAuth2Strategy as GoogleStrategy} from "passport-google-oauth";
+import {Strategy as FacebookStrategy} from "passport-facebook";
 import configs from "./config";
-import UserController from "../controllers/user";
 import {makeError} from "../utils/messenger";
 
+import UserController from "../controllers/user/user";
 
-export default function (app) {
+
+export default function login(app) {
 	const config = configs[process.env.NODE_ENV];
-	const log = debug("log:passport");
 
 	passport.serializeUser((user, callback) => {
 		callback(null, user._id);
 	});
 
-	passport.deserializeUser((id, callback) => {
+	passport.deserializeUser((_id, callback) => {
 		const userController = new UserController();
-		userController.findById(id, {lean: false})
+		userController.getByQuery({_id})
 			.then(user => {
 				callback(null, user);
 			})
@@ -33,24 +30,23 @@ export default function (app) {
 	passport.use(new LocalStrategy(config.strategies.local,
 		(email, password, callback) => {
 			const userController = new UserController();
-			userController
-				.findOne({email}, {
-					select: "+password",
-					lean: false
-				})
+			userController.findOne({email: email.toLowerCase()}, {
+				select: "+password",
+				lean: false
+			})
 				.tap(user => {
 					if (!user) {
-						throw makeError("incorrect-name", user, 401);
+						throw makeError("server.user-incorrect-name", user, 401);
 					}
 				})
 				.tap(user => {
 					if (!user.verifyPassword(password)) {
-						throw makeError("incorrect-password", user, 401);
+						throw makeError("server.user-incorrect-password", user, 401);
 					}
 				})
 				.tap(user => {
 					if (!user.isEmailVerified) {
-						throw makeError("verify-email", user, 401);
+						throw makeError("server.user-is-not-verified", user, 403);
 					}
 				})
 				.tap(user => {
@@ -62,34 +58,45 @@ export default function (app) {
 		}
 	));
 
+	function getUser(provider, profile) {
+		const userController = new UserController();
+		return userController.getByQuery({[`${provider}.id`]: profile.id})
+			.then(administrator => {
+				if (administrator) {
+					return administrator;
+				} else {
+					return userController.getByQuery({email: profile.emails[0].value.toLowerCase()})
+						.then(user => {
+							if (user) {
+								user.set(provider, profile._json);
+								return userController.save(user);
+							} else {
+								if (config.permissions.autoCreate) {
+									return userController.create({
+										fullName: `${profile.name.givenName} ${profile.name.familyName}`,
+										email: profile.emails[0].value,
+										isEmailVerified: true,
+										[provider]: profile._json
+									});
+								} else {
+									throw makeError("server.user-profile-not-found", user, 401);
+								}
+							}
+						});
+				}
+			});
+	}
+
 	if (config.strategies.google.clientID) {
 		passport.use(new GoogleStrategy(config.strategies.google,
 			(accessToken, refreshToken, profile, callback) => {
-				log("google:profile", profile);
-				const userController = new UserController();
-				userController.findOne({"google.id": profile.id}, {lean: false})
-					.then(user => {
-						if (!user) {
-							return userController
-								.create({
-									firstName: profile.name.givenName,
-									lastName: profile.name.familyName,
-									email: profile.emails[0].value,
-									isEmailVerified: true,
-									google: profile._json
-								})
-								.then(newUser => {
-									callback(null, newUser);
-								})
-								.catch(error => {
-									callback(error, null);
-								})
-								.done();
-						} else {
-							callback(null, user);
-						}
+				getUser("google", profile)
+					.tap(user => {
+						callback(null, user);
 					})
-					.catch(callback)
+					.catch(error => {
+						callback(error, null);
+					})
 					.done();
 			}
 		));
@@ -98,31 +105,13 @@ export default function (app) {
 	if (config.strategies.facebook.clientID) {
 		passport.use(new FacebookStrategy(config.strategies.facebook,
 			(accessToken, refreshToken, profile, callback) => {
-				log("facebook:profile", profile);
-				const userController = new UserController();
-				userController.findOne({"facebook.id": profile.id}, {lean: false})
-					.then(user => {
-						if (!user) {
-							return userController
-								.create({
-									firstName: profile.name.givenName,
-									lastName: profile.name.familyName,
-									email: profile.emails[0].value,
-									isEmailVerified: true,
-									facebook: profile._json
-								})
-								.then(newUser => {
-									callback(null, newUser);
-								})
-								.catch(error => {
-									callback(error, null);
-								})
-								.done();
-						} else {
-							callback(null, user);
-						}
+				getUser("facebook", profile)
+					.tap(user => {
+						callback(null, user);
 					})
-					.catch(callback)
+					.catch(error => {
+						callback(error, null);
+					})
 					.done();
 			}
 		));
