@@ -1,16 +1,19 @@
-"use strict";
-
 import session from "express-session";
+import connectRedis from "connect-redis";
 import connectMongo from "connect-mongo";
-import mongoose from "../configs/mongoose";
-import configs from "../configs/config";
+import {getConnections} from "../utils/mongoose";
+import redis from "./redis";
+import configs from "./config";
 
 
 export default function (app) {
 	const config = configs[process.env.NODE_ENV];
-	const connection = mongoose(config);
+	const connections = getConnections();
 
 	app.use(session(Object.assign({}, config.session, {
-		store: new (connectMongo(session))({mongooseConnection: connection.user})
+		store: config.session.store === "mongo" ?
+			new (connectMongo(session))({mongooseConnection: connections.auth}) :
+			new (connectRedis(session))(Object.assign(config.redis, {client: redis()}))
 	})));
 }
+
