@@ -1,84 +1,106 @@
-"use strict";
-
-import _ from "lodash";
-import path from "path";
-import RichModel from "./model";
-import DebuggableController from "./debuggable";
-import {checkModel} from "../../utils/messenger";
+import {pick, omit} from "lodash";
+import AbstractModel from "./../../models/abstract/abstract";
+import {checkModel} from "../../utils/middleware";
+import {getConnections} from "../../utils/mongoose";
 
 
-class AbstractController extends DebuggableController {
+export const fieldsToOmit = ["_id", "createdAt", "updatedAt"];
+
+export default class AbstractController {
+
+	static displayName;
 
 	static realm;
 
-	constructor(isDebuggable = true, connection) {
-		super(isDebuggable);
-		// i'm very sorry for this hack
-		const config = require(path.join(process.env.PWD, "server/configs/config")).default[process.env.NODE_ENV];
-		const mongoose = require(path.join(process.env.PWD, "server/configs/mongoose")).default(config);
-		this.constructor.realm = this.constructor.realm || config.realm;
-		this.model = new RichModel(this.constructor.displayName, isDebuggable, connection || mongoose[this.constructor.realm]);
+	static param = "_id";
+
+	constructor() {
+		this.constructor.displayName = this.constructor.name.slice(0, -10);
+		this.model = new AbstractModel(this.constructor.displayName, getConnections()[this.constructor.realm]);
+	}
+
+	getByUId(request, options) {
+		return this.findOne({
+			[this.constructor.param]: request.params[this.constructor.param] || request.body[this.constructor.param],
+			organizations: request.user.apiKey.organization
+		}, options)
+			.then(model => checkModel().bind(this)(model, request));
 	}
 
 	getById(request) {
-		return this.findOne({
-				_id: request.params._id,
-				[this.constructor.realm]: request.user._id
-			})
-			.then(checkModel(request.user).bind(this));
+		return this.getByUId(request);
 	}
 
 	list(request) {
-		return this.find({[this.constructor.realm]: request.user._id})
+		return this.find({organizations: request.user.apiKey.organization})
 			.then(list => ({list}));
 	}
 
-	insert(request, fields = []) {
+	insert(request, fieldsToPick = [], data = {}) {
 		// don't remove _id it causes recursion and `Maximum call stack size exceeded` error
-		const clean = Object.assign({}, fields.length ? _.pick(request.body, fields) : request.body, {[this.constructor.realm]: request.user._id});
+		const clean = fieldsToPick.length ? pick(request.body, fieldsToPick) : omit(request.body, fieldsToOmit);
+		Object.assign(clean, data, {organizations: [request.user.apiKey.organization._id]});
 		return this.create(clean);
 	}
 
-	edit(request, fields = []) {
-		const clean = Object.assign({}, fields.length ? _.pick(request.body, fields) : request.body);
-		return this.findOneAndUpdate({
-				_id: request.params._id,
-				[this.constructor.realm]: request.user
-			}, clean, {new: true})
-			.then(checkModel(request.user).bind(this));
+	edit(request) {
+		return this.change(request);
+	}
+
+	change(request, options = {}, conditions = [], fieldsToPick = [], data = {}) {
+		Object.assign(options, {lean: false});
+		return this.getByUId(request, options, conditions)
+			.then(item => {
+				const clean = fieldsToPick.length ? pick(request.body, fieldsToPick) : omit(request.body, fieldsToOmit);
+				if (Object.keys(clean).length || Object.keys(data).length) {
+					Object.assign(item, clean, data);
+					return this.save(item);
+				} else {
+					return item;
+				}
+			});
+	}
+
+	deactivate(request) {
+		return this.getByUId(request)
+			.then(model =>
+				this.remove(model)
+			);
 	}
 
 	delete(request) {
-		return this.findOneAndRemove({
-				_id: request.params._id,
-				[this.constructor.realm]: request.user
-			})
-			.then(checkModel(request.user).bind(this))
+		return this.deactivate(request)
 			.thenResolve({success: true});
+	}
+
+	conditions(request, conditions = []) {
+		return item => {
+			conditions.forEach(condition => condition.bind(this)(item, request));
+			return item;
+		};
 	}
 
 }
 ; // eslint-disable-line no-extra-semi
 
 [
+	"aggregate",
 	"count",
 	"distinct",
-	"remove",
-	"destroy",
 	"create",
-	"aggregate",
-	"mapReduce",
 	"find",
-	"findOne",
 	"findById",
 	"findByIdAndRemove",
 	"findByIdAndUpdate",
+	"findOne",
 	"findOneAndRemove",
 	"findOneAndUpdate",
+	"fulltextsearch",
+	"mapReduce",
 	"populate",
-	"update",
-	"search",
+	"remove",
 	"save",
+	"update",
 	"upsert"
 ].forEach(name => {
 	AbstractController.prototype[name] = function (...args) {
@@ -86,4 +108,3 @@ class AbstractController extends DebuggableController {
 	};
 });
 
-export default AbstractController;
