@@ -1,9 +1,10 @@
-"use strict";
-
 import {Schema} from "mongoose";
 
 const Avatar = new Schema({
 	url: String
-}, {versionKey: false});
+}, {
+	timestamps: true,
+	versionKey: false
+});
 
 export default Avatar;
