@@ -1,9 +1,6 @@
-"use strict";
-
-
 export const email = "trejgun@gmail.com";
 export const password = "My5up3r5tr0ngP@55w0rd";
 export const confirm = password;
-export const firstName = "Trej";
-export const lastName = "Gun";
+export const fullName = "Trej Gun";
 export const phoneNumber = "+62 (812) 3919-8760";
+export const goatId = "ffffffffffffffffffffffff";
