@@ -1,69 +1,67 @@
-"use strict";
+import winston from "winston";
+import {makeError} from "./messenger";
+import {translate} from "./lang";
 
-import debug from "debug";
-import {makeError} from "../utils/messenger";
-import {translate} from "../utils/lang";
-
-const log = debug("utils:response");
 
 function _send(request, response) {
 	return (error) => {
-		response.status(error.code).send({
-			code: error.code,
+		response.status(error.status).send({
+			status: error.status,
 			errors: [].concat(error.message)
 		});
 	};
 }
 
+export function processValidationError(error) {
+	return Object.keys(error.errors).map(key => error.errors[key].reason || error.errors[key].message);
+}
+
+export function processMongoError(error, user) {
+	const key = `mongo.${error.message.match(/\$(\S+)/)[1]}`;
+	const translation = translate(key, user);
+	return translation === key ? translate("mongo.E11000", user) : translation;
+}
+
 export function sendError(error, request, response, next) {
-	log("sendError", error);
+	winston.error("->", error);
 	void next; // eslint
 	const send = _send(request, response);
+
 	if (error.name === "ValidationError") {
 		return send({
-			code: 409,
-			message: Object.keys(error.errors).map(key => error.errors[key].message)
+			status: 409,
+			message: processValidationError(error)
 		});
 	}
+
 	if (error.name === "MongoError" && error.code === 11000) {
-		const key = error.message.match(/\$(\S+)/)[1];
 		return send({
-			code: 400,
-			message: translate("error/mongo/" + key, request.user) || translate("error/mongo/E11000", request.user)
+			status: 400,
+			message: processMongoError(error, request.user)
 		});
 	}
-	if (error.type === "StripeCardError" || error.type === "StripeInvalidRequest") {
-		return send({
-			code: 400,
-			message: error.message
-		});
-	}
-	if (!error.code) {
+
+	if (!error.status) {
 		if (process.env.NODE_ENV === "production") {
-			return send(makeError("server-error", request.user, 500));
+			return send(makeError("server.error", request.user, 500));
 		} else {
 			return send({
-				code: 500,
+				status: 500,
 				message: error.stack
 			});
 		}
 	}
+
 	return send(error);
 }
 
 export function wrapJSON(method) {
 	return (request, response, next) => {
 		method(request, response, next)
-			.then(result => {
-				if (result.success === false) {
-					response.status(400);
-				}
-				response.json(result);
-			})
-			.catch(error => {
-				log(error);
-				return sendError(error, request, response);
-			})
+			.then(::response.json)
+			.catch(error =>
+				sendError(error, request, response)
+			)
 			.done();
 	};
 }
@@ -71,13 +69,9 @@ export function wrapJSON(method) {
 export function wrapFile(method) {
 	return (request, response, next) => {
 		method(request, response, next)
-			.then(result => {
-				log("result", result);
-			})
-			.catch(error => {
-				log(error);
-				return sendError(error, request, response);
-			})
+			.catch(error =>
+				sendError(error, request, response)
+			)
 			.done();
 	};
 }
