@@ -1,92 +1,93 @@
-"use strict";
+import {makeError} from "./error";
+import {intersection} from "lodash";
+import {date} from "../utils/constants/date";
 
 
-import {isType} from "./utils";
-import {makeError} from "../utils/messenger";
-
-const validator = {
-	type(val, type) {
-		// because value is actually a string
-		return type === Number ? Number(val).valueOf().toString() === val : isType(val, type.name);
-	},
-	min(val, min) {
-		return val >= min;
-	},
-	max(val, max) {
-		return val <= max;
-	},
-	minlength(val, min) {
-		return val.length >= min;
-	},
-	maxlength(val, max) {
-		return val.length <= max;
-	},
-	enum(val, options) {
-		return !~options.indexOf(val);
-	},
-	required(val) {
-		return !!val;
-	},
-	regexp(val, regexp) {
-		return regexp.test(val);
-	},
-	default() {
-		return true; // always true
-	}
-};
-
-export function requiresLogin(request, response, next) {
-	if (!request.isAuthenticated()) {
-		return next(makeError("login", null, 401));
-	}
-	return next();
-}
-
-export function requiresRole(required) {
-	return (request, response, next) => {
-		requiresLogin(request, response, () => {
-			if (!~required.indexOf(request.user.role)) {
-				return next(makeError("access-denied", request.user, 403));
-			}
-			return next();
-		});
-	};
-}
-
-export function methodNotAllowed(request, response, next) {
-	return next(makeError("method-not-allowed", request.user, 405));
-}
-
-
-export function validateParams(rules, source) {
-	return (request, response, next) => {
-		const errors = [];
-		Object.keys(rules).forEach(name => {
-			if (!(name in request[source])) {
-				if ("default" in rules[name]) {
-					Object.assign(request[source], {[name]: rules[name].default});
-				} else if (rules[name].required === true) {
-					errors.push(makeError("no-param", request.user));
-				}
-			} else if (!Object.keys(rules[name]).every(key => validator[key](request[source][name], rules[name][key]))) {
-				errors.push(makeError("invalid-param", request.user));
-			}
-		});
-		return next(errors[0]);
-	};
-}
-
-export function validatePagination() {
-	return validateParams({
-		skip: {
-			type: Number,
-			default: 0
-		},
-		limit: {
-			type: Number,
-			default: 50,
-			min: 0,
-			max: 100
+export function checkActive(isAllowed = false, statuses = ["active"]) {
+	return function checkActiveInner(model, request) {
+		const isAdmin = !request.user.apiKey.public;
+		const isActive = statuses.includes(model.status);
+		if (!isActive && !(isAllowed && isAdmin)) {
+			throw makeError(`not-active.${this.constructor.displayName.toLowerCase()}`, request.user, 400);
+		} else {
+			return model;
 		}
-	}, "query");
+	};
 }
+
+export function checkModel() {
+	return function checkModelInner(model, request) {
+		if (!model) {
+			throw makeError(`not-found.${this.constructor.displayName.toLowerCase()}`, request.user, 404);
+		} else {
+			return model;
+		}
+	};
+}
+
+export function checkOwner(isAllowed = false) {
+	return function checkOwnerInner(model, request) {
+		const u = request.user.apiKeys.map(apiKey => apiKey.organizations[0]._id.toString());
+		if (u.includes(model.owner.toString()) === isAllowed) {
+			throw makeError("server.access-denied", request.user, 403);
+		} else {
+			return model;
+		}
+	};
+}
+
+export function checkPast(isAllowed = false, field = "startTime") {
+	return function checkPastInner(model, request) {
+		const isAdmin = !request.user.apiKey.public;
+		const isPast = model[field] <= date; // <= for tests
+		if (isPast && !(isAllowed && isAdmin)) {
+			throw makeError("controller.event-has-passed", request.user);
+		} else {
+			return model;
+		}
+	};
+}
+
+export function checkPublic() {
+	return function checkPublicInner(model, request) {
+		if (!model.public) {
+			throw makeError("server.access-denied", request.user, 403);
+		} else {
+			return model;
+		}
+	};
+}
+
+export function checkPublicContract(publicContract) {
+	return function checkPublicContractInner(model, request) {
+		const isPublic = request.params._id.toString() === publicContract._id.toString();
+		if (isPublic && !request.body.public) {
+			throw makeError("controller.contract-must-have-public-contract", 400);
+		} else {
+			return model;
+		}
+	};
+}
+
+export function checkOrganization(isAllowed = false) {
+	return function checkOrganizationInner(model, request) {
+		const m = model.organizations.map(organization => organization._id.toString());
+		const u = request.user.apiKeys.map(apiKey => apiKey.organizations[0]._id.toString());
+		if (intersection(m, u).length > 0 === isAllowed) {
+			throw makeError("server.access-denied", request.user, 403);
+		} else {
+			return model;
+		}
+	};
+}
+
+export function checkPermissions(permission, realm) {
+	return function checkPermissionsInner(model, request) {
+		if (!request.user.apiKey.permissions[realm || this.constructor.realm].includes(permission)) {
+			throw makeError("server.access-denied", request.user, 403);
+		} else {
+			return model;
+		}
+	};
+}
+
